Release 232.0.0 (#4861)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

Recently, a new version of `@metamask/network-controller`, 22.0.0, was
released, but packages which have a peer dependency on
`@metamask/network-controller` were not included in this release. This
means that clients which are using current versions of these dependents
will experience peer dependency warnings if they attempt to upgrade
`@metamask/network-controller`. This release makes it so that clients
can also upgrade these dependents while they upgrade
`@metamask/network-controller`, avoiding the peer dependency warnings.

## References

This is a prequisite for
https://github.com/MetaMask/MetaMask-planning/issues/3568.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes