chore: Move BigNumber.js to dependencies (#4873)
## Explanation

Since `bignumber.js` is used for typing in `controller-utils`, it should
likely be a dependency to prevent issues when building with TS
downstream.

There is probably a larger conversation to have here about when to use
types as devDeps and when to use them as deps, but in our architecture
today we rely heavily on all types to be included as dependencies. This
causes a lot of potentially avoidable dependency management.

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/controller-utils`

- **Fixed**: Move `bignumber.js` to dependencies

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes