chore: add PRs to PR queue for WF team (#4891)
## Explanation

This GitHub Action automatically adds pull requests to a project board
for the Wallet Framework Team. It triggers on PR events like opening,
labeling, or review requests. The action checks if the PR is relevant by
matching the requested team or labels against the Wallet Frameworks team
name and label. If matched, it adds the PR to our project board using a
GitHub token for authentication. This will allow the team to more easily
see PRs which need our attention

## References
n/a

## Changelog
n/a

## Testing
I did manual testing on a fork of the repo. It is not a one for one
match because the access token setup is slightly different based on
Organization -> projects vs Individual -> projects but the core of the
workflow worked and is the same.

## Checklist (none applicable) 

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes