Improve documentation for contributors/maintainers
It was pointed out recently that the link to the Contributing guide is
buried at the bottom of the README. This commit makes it more prominent
by moving it to the top and lists the major sections of the document.

However, this commit also updates the Contributing guide to be more
generally helpful and to answer questions that I've received from other
teams not familiar with this repo who are unaware that codeownership and
releases are shared among teams.