Add script to generate peer dependency graph
When upgrading controller packages in clients, some packages need to be
upgraded in tandem with others to satisfy peer dependencies. This commit
adds a new script that can be used to view all of the workspaces in the
monorepo along with the internal peer dependencies that each workspace
may rely upon.

For instance, if you run

```
yarn run workspaces:list-peer-dependencies
```

you will get the following command output:

```
├╴ @metamask/accounts-controller 19.0.0
│  └╴ @metamask/keyring-controller ^18.0.0
├╴ @metamask/address-book-controller 6.0.1
├╴ @metamask/announcement-controller 7.0.1
├╴ @metamask/approval-controller 7.1.1
├╴ @metamask/assets-controllers 44.1.0
│  ├╴ @metamask/accounts-controller ^19.0.0
│  ├╴ @metamask/approval-controller ^7.0.0
│  ├╴ @metamask/keyring-controller ^18.0.0
│  ├╴ @metamask/network-controller ^22.0.0
│  └╴ @metamask/preferences-controller ^14.0.0
├╴ @metamask/base-controller 7.0.2
├╴ @metamask/build-utils 3.0.1
...
```