Fix peer dependency warnings
When running `yarn install`, several packages generate peer dependency
warnings. `@lavamoat/preinstall-always-fail` is now a peer dependency of
`@lavamoat/allow-scripts`, so that makes sense, but for the other
packages (`@babel/runtime`, `@metamask/providers`,
`webextension-polyfill`), we really all of these packages to present as
dependencies of the _client_, so they just need to be added as dev
dependencies here.