Fix peer dependency warnings (#4974)
When running `yarn install`, several packages generate peer dependency
warnings.

- `@lavamoat/preinstall-always-fail` is now a peer dependency of
`@lavamoat/allow-scripts`, so we need to add it as a new dev dependency.
- For the other packages — `@babel/runtime`, `@metamask/providers`,
`webextension-polyfill` — we really expect the _client_ to have these
dependencies, so we also add them as dev and peer dependencies instead
of dependencies. We have to do this because they will not get forwarded
along.