Only enable Node lint rules for Node files (#3672)
In a previous commit, we mistakenly introduced a Node-specific function
for testing deep equality, and this ended up crashing the extension.
This would usually have been caught by our ESLint rules, as they
prohibit use of Node libraries by default. However, the ESLint
configuration for this repo imports rules from our
`@metamask/eslint-config-nodejs` package and applies them to all files,
marking everything as Node-compatible, thereby allowing such usage. This
is incorrect: these rules should only apply to test files and scripts.

This commit fix the ESLint configuration to match. Doing so revealed a
couple of categories of violations, which this commit also fixes:

- Some packages import and make use of EventEmitter from Node's `events`
module. We add a notice to the README for these packages which advises
consumers that these packages are designed to be used in a
Node-compatible environment.
- Some packages import and make use of Node's `assert` module to check
values at runtime. It isn't strictly necessary to use this module, and
so we replace its usage with simpler code.
- Some packages import and make use of Node's `inspect` utility
function. We don't strictly need this either, and we replace its usage
with simpler code as well.