Don't export Allowed{Actions,Events} in example controllers (#4996)
## Explanation

Our controller guidelines states that the `AllowedActions` and
`AllowedEvents` types in controller files should not be exported. Our
example controllers do not follow this guideline, so this commit
corrects this.

## References

(None)

## Changelog

None (change to contributor documentation only)

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes