Add label for use in github project auto add workflow (#5042)
## Explanation

We have a github project setup to track client controller upgrade issues
that are created by this github action. Github projects workflow is not
working for auto adding solely based on title match and the options for
what to match on are very limited. Labels are one way we can do that.

This PR adds a label that we can trigger the auto add workflow on. 

## References


## Changelog

n/a

## Checklist
n/a