Use common function for service policy boilerplate
We would like to use the Cockatiel library in our service classes to
ensure that requests are retried using the circuit breaker pattern. Some
of our service classes do this already, but we are copying and pasting
the code around. This commit extracts the boilerplate code to a new
function in the `@metamask/controller-utils` package,
`createServicePolicy`, so that we no longer have to do this.