Start extracting service policy boilerplate to common function
We would like to use the Cockatiel library in our service classes to
ensure that requests are retried using the circuit breaker pattern. Some
of our service classes do this already, but we are copying and pasting
the code around. This commit extracts some of the boilerplate code to a
new function in the `@metamask/controller-utils` package,
`createServicePolicy` so that we no longer have to do this. Because it
would require an enormous amount of tests, this commit does not include
a complete version of `createServicePolicy`, notably omitting
`maxRetries` option as well as a way to handle slow requests; those will
be added later.