chore: Rename `ControllerMessenger` to `Messenger`
Rename the `ControllerMessenger` to `Messenger` so that it is clear it
can be used for more than just controllers.

This was decided by this ADR: https://github.com/MetaMask/decisions/blob/main/decisions/core/0001-messaging-non-controllers.md

Previous names have been preserved as aliases to avoid making this a breaking
change.

Relates to #4538