fix: check push notification compatibility before setting up firebase (#5069)
## Explanation

Before we were initialising firebase without checking if the browser has
correct setup for it. Thus causing errors logs to be made. (not breaking
the app, but still producing errors).

We now do a check to ensure we can initialise firebase.

## References

https://github.com/MetaMask/metamask-extension/issues/26173

## Changelog

### `@metamask/notification-services-controller`

- **ADDED**: `getPushAvailability` method to check if we can use push
notifications.
- **CHANGED**: `getFirebaseMessaging()` now optionally returns null.
This is an internal method not exposed outside.
- **CHANGED**: `unsubscribePushNotifications` is not optional. This is
internal and not exposed outside module.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes