Bump `@metamask/utils` to `^11.0.1` and `@metamask/rpc-errors` to `^7.0.2` (#5080)
## Explanation

This bumps `@metamask/utils` from `^10.0.0` to `^11.0.1`, in an effort
to bump all packages using `@metamask/utils` to use v11. Unfortunately
there are some circular dependencies, so this does not completely
eliminate previous versions yet, but this is required to bump
`@metamask/utils` in other packages, such as:

- https://github.com/MetaMask/eth-json-rpc-middleware/pull/351

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

- **CHANGED**: Bump `@metamask/utils` from `^10.0.0` to `^11.0.1`
- **CHANGED**: Bump `@metamask/rpc-errors` from `^7.0.1` to `^7.0.2`

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes