fix: add notification call prevention logic (#5081)
## Explanation

We've added logic on platforms to make sure that the global
notifications switch is enabled before we fetch. But ideally this should
be internal logic.

So we've updated the logic to ensure that Feature Announcements and
Wallet Notifications will bail early if the toggle is off. Snap
notifications however will not obey this flag (to keep compatibility
with the original snap notifications feature).

## References

https://github.com/MetaMask/metamask-extension/issues/28173

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/notification-services-controller`

- **CHANGED**: add `isNotificationServicesEnabled` check inside
`fetchAndUpdateMetamaskNotifications` method.
- this will prevent fetching feature announcements and wallet
notifications if set to `false`.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes