Gracefully handle requesting gas fee estimates for nonexistent network
Currently, when GasFeeController tries to fetch gas fee estimates for a
network that does not exist for some reason, it consults
NetworkController for the corresponding network client and then
NetworkController throws an error.

It could be argued that we don't really care that the network doesn't
exist and there is no reason to expose this information to the client.
Given this assumption, this commit gracefully handles this situation by
returning an empty set of estimates instead. It also extracts the error
message produced when a network client is not found so that it can be
checked more easily.