Transition to new process for updating changelogs
Currently, we ask that when contributors submit a pull request, they
fill out a Changelog section in the description which lists all of the
consumer-facing changes they've made. This section is ultimately used
when creating a new release to assist with populating `CHANGELOG.md`
files.

However, this process is difficult to follow in practice because it
forces release authors to do work on behalf of other engineers. The
people that know how to describe changes best are those that made the
change, so they should be reponsible for updating changelogs.

With this commit, we now direct contributors to update changelog files
directly as they make changes instead of placing them in the Changelog
section of the PR description (which no longer exists). This commit adds
documentation on how to do this effectively (much of which is now in the
`contributor-docs` repo).