Add NetworkController RPC service class
This commit adds an RPC service class that will ultimately be used
inside of middleware to make JSON-RPC requests to an RPC endpoint. It
makes uses of `createServicePolicy`, added in a previous commit, to
encapsulate the request code using the retry and circuit breaker
policies. As this service class is designed to replace large parts of
the fetch and Infura middleware, it customizes the service policy so
that the request will be retried only when the network is perceived to
be "down". This occurs when:

- The `fetch` call throws a "Failed to fetch" error (or something
  similar to it; see code for the full list of variations)
- The `fetch` call throws an ETIMEDOUT or ECONNRESET error
- The response status is 503 or 504
- The response body is invalid JSON

In contrast, the network is not perceived to be "down" if:

- The `fetch` call throws an unexpected error (e.g. if the request
  options are invalid)
- The response status is not 2xx, but is also not 503 or 504
- The response body is an unsuccessful JSON-RPC response
- The response body is a successful, but empty, JSON-RPC response