Fix ESLint config (#5132)
The upgrade to ESLint 9 which occurred in a previous commit was not
performed correctly. It seems that an array must be passed to
`createConfig` instead of a series of arguments, and that was not done.
As a result, a lot of lint rules were disabled accidentally.

This commit fixes the ESLint config and corrects any lint violations.
Unsurprisingly, upgrading the ESLint packages created a bunch of new
lint violations, so those have been turned into warnings rather than
errors in this commit. To prevent new instances of these violations from
occurring, however, a script has been added which acts as a quality
gate, running `eslint` as before, but then capturing existing warning
counts and comparing them to a previously generated threshold file,
exiting with non-zero if there is an increase in warnings for any
particular ESLint rule. This script is also now used for `yarn
lint:eslint` so that CI will fail if the count increases.