Relax switch-exhaustiveness-check lint rule (#5157)
The default behavior for this rule enforces that a switch statement
handle all possible values that the given variable could be, including
type unions. This rule seems to be acting differently in different
environments, especially CI, preventing developers from merging PRs. We
could disable this rule entirely, but let's try relaxing it a bit to see
if that improves the situation.