test(accounts-controller): improve uuid mocks (#5005)
## Explanation

Unify the UUID mocking. With this new refactor we longer have to specify
each returned values of `mockUUID`. It also handles the non-registered
accounts (in this case, a random UUID will be generated).

## References

N/A

## Changelog

N/A (those changes only impact tests)

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes