fix: add thresshold for update NFT metadata (#5134)
## Explanation

When `onPreferencesControllerStateChange` is triggered inside
nftController, it also triggers a
[check](https://github.com/MetaMask/core/blob/76af6ebde290a270f4da6493f0f3b6c9354545cb/packages/assets-controllers/src/NftController.ts#L431)
if nfts saved in state needs to be updated.
We keep retrying to update NFTs (fetching their metadata) when they have
no image, name or description.

Note: When switching newtworks; this
[line](https://github.com/MetaMask/metamask-extension/blob/b91a9622393f37b79a63ba4a240cda7efa3cc5bc/ui/components/multichain/network-list-menu/network-list-menu.tsx#L296)
is the trigger for the `onPreferencesControllerStateChange`.

While the updateNftUpdateForAccount should be triggered once for NFTs
that actually have a (name, description, image)
We will keep retrying for NFTs that don't. (Could be nfts created for
testing purposes or some raffle tickets do not have those fields).

We had a user reporting that he is seeing an excessive amount of RPC
calls made on one of his accounts on Arbitrum Sepolia. This account
holds more than a 1000 NFT. Every time he switches the network to
Arbitrum Sepolia, he sees more than a 1000 eth_call made.

That was the result of updateNftUpdateForAccount being triggered
everytime the network is switched + The nfts not having (name,image,
description)


This PR adds a 500 threshold to avoid sending an excessive amount of RPC
calls to fetch nft metadata and adds a check if the triggered
onPreferencesControllerStateChange updates any of the concerned params (
ipfsGateway, openseaEnabled, isIpfsGatewayEnabled)


**_Better improvement:_**

A better improvement for this would be to add support for
[BalanceScanner](https://github.com/MetaMask/metafi-sdk/blob/main/packages/eth-scan/contracts/BalanceScanner.sol)
contract, which allows to fetch tokenURI for multiple addresses and
tokenIds with fct tokenURIsofTokenIdsForTokens.

We will control how often this is refreshed. For users that have nfts >
NFT_UPDATE_THRESHOLD, we save the timestamp of the first update in a
mapping chainId => lastTimestamp of updateNftMetadata, and only update
after a specific period of time NFT_UPDATE_REFRESH_PERIOD = 48hours.


## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/controller-utils`

- **ADDED**: Added NFT_UPDATE_THRESHOLD constant


### `@metamask/assets-controllers`

- **CHANGED**: Added a check if the nftsToUpdate is less than
NFT_UPDATE_THRESHOLD inside `updateNftUpdateForAccount`

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: Michele Esposito <34438276+mikesposito@users.noreply.github.com>